spike of adding some notion of linking input to output
