+++ parser-typechecker/src/Unison/Codebase/Editor/Command.hs
@@ -26,7 +26,8 @@ import           Unison.Server.Backend          ( DefinitionResults
 import           Data.Configurator.Types        ( Configured )
 import qualified Data.Map                      as Map
 
-import           Unison.Codebase.Editor.Output
+-- import           Unison.Codebase.Editor.Output
+import           Unison.Codebase.Editor.InputOutput
 import           Unison.Codebase.Editor.RemoteRepo
 
 import           Unison.Codebase.Branch         ( Branch )
@@ -64,8 +65,8 @@ import qualified Unison.WatchKind as WK
 import Unison.Codebase.Type (GitError)
 
 type AmbientAbilities v = [Type v Ann]
-type SourceName = Text
-type Source = Text
+-- type SourceName = Text
+-- type Source = Text
 type LexedSource = (Text, [L.Token L.Lexeme])
 
 data LoadSourceResult = InvalidSourceNameError
@@ -104,6 +105,8 @@ data Command m i v a where
   ConfigLookup :: Configured a => Text -> Command m i v (Maybe a)
 
   Input :: Command m i v i
+
+  InputWithOutput :: Command m i v i -> Output v -> Command m i v i
 
   -- Presents some output to the user
   Notify :: Output v -> Command m i v ()
@@ -253,6 +256,7 @@ commandName = \case
   UI                          -> "UI"
   ConfigLookup{}              -> "ConfigLookup"
   Input                       -> "Input"
+  InputWithOutput{}           -> "Input"
   Notify{}                    -> "Notify"
   NotifyNumbered{}            -> "NotifyNumbered"
   AddDefsToCodebase{}         -> "AddDefsToCodebase"
