+++ parser-typechecker/src/Unison/Codebase/Editor/HandleInput.hs
@@ -28,9 +28,9 @@ import Unison.Server.Backend (ShallowListEntry(..), TermEntry(..), TypeEntry(..)
 import qualified Unison.Codebase.MainTerm as MainTerm
 import Unison.Codebase.Editor.Command as Command
 import Unison.Codebase.Editor.Input
-import Unison.Codebase.Editor.Output
+import Unison.Codebase.Editor.InputOutput
 import Unison.Codebase.Editor.DisplayObject
-import qualified Unison.Codebase.Editor.Output as Output
+import qualified Unison.Codebase.Editor.InputOutput as Output
 import Unison.Codebase.Editor.SlurpResult (SlurpResult(..))
 import qualified Unison.Codebase.Editor.SlurpResult as Slurp
 import Unison.Codebase.Editor.SlurpComponent (SlurpComponent(..))
